Convert SimpleSocketClient into 'reactive' api
This is in preparation of making RSocktuatorClient (using RSocket)
and SimpleSocketClient implement the same SocktuatorClient interface.

Sadly this 'common' api will have to be reactive because it isn't advisable to take
a reactive api and wrap it in something that calls '.block()' everywhere. This usually results in somewhat rando deadlocks when calls end up going full circle between reactive -> blocking -> reactive.